requirements.txt
census==0.8.15
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
jellyfish==0.6.1
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.10
us==2.0.2
